% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bettergetwd.R
\name{bettergetwd}
\alias{bettergetwd}
\title{Better getwd()}
\usage{
bettergetwd()
}
\value{
An address with all the escape slashes already added for you.
}
\description{
Better getwd()
}
\examples{
\dontrun{mygetwd()}

}
