% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stan.R
\name{fit_stan}
\alias{fit_stan}
\title{Fit Stan model to data, return stanfit object}
\usage{
function(y, x=NA, model_name = NA, est_drift = FALSE, P = 1, Q = 1, mcmc_list = list(n_mcmc = 1000, n_burn = 500, n_chain = 3, n_thin = 1), ...)
}
\arguments{
\item{y}{The response data}

\item{x}{An optional vector or matrix (regression only) of data}

\item{model_name}{The type of model to be fit; options include 'regression', 'ar', 'rw', 'ma', 'ss_ar' (state space univariate AR), or 'ss_rw' (state space univariate random walk)}

\item{est_drift}{Whether or not the mean or drift term is estimated (not applicable for regression)}

\item{P}{Degreee of AR process, minimum = 1 (default)}

\item{Q}{Degreee of MA process, minimum = 1 (default)}

\item{mcmc_list}{List controlling the MCMC parameters; includes n_mcmc (chain length, defaults to 1000), n_burn (burn-in, defaults to 500), n_chain = 3 (number of chains),
n_thin = 1 (thinning rate)}

}
\description{
Fit a Stan model and return output for all parameters, including predicted values.
}

