% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{My Confidence Interval}
\usage{
myci(data, conf.level = 0.95)
}
\arguments{
\item{data}{data}

\item{conf.level}{conf.level}
}
\value{
Confidence Interval
}
\description{
My Confidence Interval
}
