% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.R
\name{myclt}
\alias{myclt}
\title{Central Limit Theorem}
\usage{
myclt(n, iter, a = 0, b = 1)
}
\arguments{
\item{n}{Number of observations per iteration}

\item{iter}{Number of iterations}

\item{a}{Lower bound of uniform distribution}

\item{b}{Upper bound of uniform distribution}
}
\value{
Histogram of the sum of observations across iterations
}
\description{
Central Limit Theorem
}
