% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncruve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{The mean}

\item{sigma}{The standard deviation}

\item{a}{Value for area calculation}
}
\value{
The area
}
\description{
myncurve
}
