% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mytt.R
\name{myttest}
\alias{myttest}
\title{mytt}
\usage{
myttest(x, y, mu = 0, paired = FALSE, conf.level = 0.95)
}
\arguments{
\item{x}{}

\item{y}{}

\item{mu}{}

\item{paired}{}

\item{conf.level}{}
}
\description{
mytt
}
