% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{Number of tickets sold}

\item{gamma}{True probability of overbooking for the flight}

\item{p}{Probability of a show}
}
\value{
Values and discrete and continuous graphs
}
\description{
ntickets
}
