\name{stats-package}
\alias{stats-package}
\alias{stats}
\docType{package}
\title{
The R Stats Package
}
\description{
R statistical functions
}
\details{
This package contains functions for statistical calculations
and random number generation.

For a complete
list of functions, use \code{library(help="stats")}.

}
\author{
R Development Core Team and contributors worldwide

Maintainer: R Core Team \email{R-core@r-project.org}
}
\keyword{ package }
