% File src/library/stats/man/ts-defunct.Rd
% Part of the R package, http://www.R-project.org
% Copyright 1995-2007 R Core Team
% Distributed under GPL 2 or later

\name{ts-defunct}
\alias{arima0.diag}
\title{Defunct functions in Package \pkg{ts}}
\description{
  The functions or variables listed here are no longer part of \R as
  they are not needed (any more).
}
\usage{
% Removed in 1.6.0
arima0.diag(\dots)
% </Removed 1.6.0>
}
\details{
  \code{arima0.diag} has been replaced by \code{\link{tsdiag.arima0}}.
}
\keyword{ts}
\keyword{internal}
