% File src/library/stats/man/stats-deprecated.Rd
% Part of the R package, https://www.R-project.org
% Copyright 1995-2014, 2020 R Core Team
% Distributed under GPL 2 or later

\name{stats-deprecated}
\alias{stats-deprecated}
%------ NOTE:  ../R/stats-deprecated.R   must be synchronized with this!
\title{Deprecated Functions in Package \pkg{stats}}
%------ PLEASE: one \alias{.} for EACH ! (+ one \usage{} & \arguments{} for all)
\description{
  These functions are provided for compatibility with older versions of
  \R only, and may be defunct as soon as the next release.
}
%% \usage{
%% plclust(tree, hang = 0.1, unit = FALSE, level = FALSE, hmin = 0,
%%         square = TRUE, labels = NULL, plot. = TRUE,
%%         axes = TRUE, frame.plot = FALSE, ann = TRUE,
%%         main = "", sub = NULL, xlab = NULL, ylab = "Height")
%% }
%% \arguments{
%%   \item{tree}{an object of the type produced by \code{\link{hclust}}.}
%%
%%  .....
%%  .....
%% }
\details{
  %-- Either:
  There are currently no deprecated functions in this package.
  %-- Or:
  %% explain *why* it's deprecated, and \code{\link{..}} to new
  %% \code{plcust} is a deprecated wrapper for the \code{plot} method for
  %% \code{\link{hclust}}, provided long ago for S-PLUS compatibility.
}
\seealso{
  \code{\link{Deprecated}}
}
\keyword{misc}
