\name{stats-deprecated}
\alias{stats-deprecated}
%------ NOTE:  ../R/Deprecated.R   must be synchronized with this!
\title{Deprecated Functions in Stats package}
%------ PLEASE: put \alias{.} here for EACH !
\description{
  These functions are provided for compatibility with older versions of
  \R only, and may be defunct as soon as the next release.
}
\details{  
  There are currently no deprecated functions in this package.
}
\seealso{
  \code{\link{Deprecated}}
}
\keyword{misc}
