% File src/library/stats/man/stats-package.Rd
% Part of the R package, http://www.R-project.org
% Copyright 1995-2007 R Core Team
% Distributed under GPL 2 or later

\name{stats-package}
\alias{stats-package}
\alias{stats}
\docType{package}
\title{
The R Stats Package
}
\description{
R statistical functions
}
\details{
This package contains functions for statistical calculations
and random number generation.

For a complete
list of functions, use \code{library(help = "stats")}.

}
\author{
R Core Team and contributors worldwide

Maintainer: R Core Team \email{R-core@r-project.org}
}
\keyword{ package }
