\name{ts-defunct}
\alias{arima0.diag}
\title{Defunct functions in Package ts}
\description{
  The functions or variables listed here are no longer part of \R as
  they are not needed (any more).
}
\usage{
% Removed in 1.6.0
arima0.diag(\dots)
% </Removed 1.6.0>
}
\details{
  \code{arima0.diag} has been replaced by \code{\link{tsdiag.arima0}}.
}
\keyword{ts}
\keyword{internal}
